+++ packages/react-reconciler/src/ReactTypeOfMode.js
@@ -9,11 +9,11 @@
 
 export type TypeOfMode = number;
 
-export const NoMode = 0b00000;
-export const StrictMode = 0b00001;
-// TODO: Remove BlockingMode and ConcurrentMode by reading from the root
-// tag instead
-export const BlockingMode = 0b00010;
-export const ConcurrentMode = 0b00100;
-export const ProfileMode = 0b01000;
-export const DebugTracingMode = 0b10000;
+export const NoMode = /*            */ 0b000000;
+// TODO: Remove BlockingMode and ConcurrentMode by reading from the root tag instead
+export const BlockingMode = /*      */ 0b000001;
+export const ConcurrentMode = /*    */ 0b000010;
+export const ProfileMode = /*       */ 0b000100;
+export const DebugTracingMode = /*  */ 0b001000;
+export const StrictLegacyMode = /*  */ 0b010000;
+export const StrictEffectsMode = /* */ 0b100000;
