+++ packages/shared/forks/ReactFeatureFlags.www.js
@@ -27,7 +27,8 @@ export const {
   decoupleUpdatePriorityFromScheduler,
   enableDebugTracing,
   skipUnmountedBoundaries,
-  enableDoubleInvokingEffects,
+  enableStrictEffects,
+  createRootStrictEffectsByDefault,
   enableUseRefAccessWarning,
   disableNativeComponentFrames,
   disableSchedulerTimeoutInWorkLoop,
